third_party/blink/web_tests/platform/generic/storage/indexeddb/mozilla/delete-result-expected.txt

Test IndexedDB: result property after deleting existing and non-existing record

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

dbname = "delete-result.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
objectStore = db.createObjectStore('foo', { keyPath: 'id', autoIncrement: true });
request = objectStore.add({});
id = event.target.result;
request = objectStore.delete(id);
PASS event.target.result is undefined
request = objectStore.delete(id);
PASS event.target.result is undefined
PASS successfullyParsed is true

TEST COMPLETE
